Refactor auth package - merge internal/microsoft
